refactor: change batch quantity to double

+++ inventory-service/src/main/java/com/ttdat/inventoryservice/application/commands/batch/UpdateProductBatchCommand.java
@@ -24,7 +24,7 @@ public class UpdateProductBatchCommand {
 
     LocalDate receivedDate;
 
-    Integer quantity;
+    Double quantity;
 
     List<CmdProductBatchLocation> batchLocations;
 }
